Move VideoId to the shared mooc module and move VideoUrl to videos module
