+++ src/Mooc/Shared/Domain/Videos/VideoId.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace CodelyTv\Mooc\Videos\Domain;
+namespace CodelyTv\Mooc\Shared\Domain\Videos;
 
 use CodelyTv\Shared\Domain\ValueObject\Uuid;
