+++ src/Mooc/Videos/Application/Create/VideoCreator.php
@@ -5,12 +5,12 @@ declare(strict_types=1);
 namespace CodelyTv\Mooc\Videos\Application\Create;
 
 use CodelyTv\Mooc\Shared\Domain\Courses\CourseId;
-use CodelyTv\Mooc\Shared\Domain\Videos\VideoUrl;
+use CodelyTv\Mooc\Shared\Domain\Videos\VideoId;
 use CodelyTv\Mooc\Videos\Domain\Video;
-use CodelyTv\Mooc\Videos\Domain\VideoId;
 use CodelyTv\Mooc\Videos\Domain\VideoRepository;
 use CodelyTv\Mooc\Videos\Domain\VideoTitle;
 use CodelyTv\Mooc\Videos\Domain\VideoType;
+use CodelyTv\Mooc\Videos\Domain\VideoUrl;
 use CodelyTv\Shared\Domain\Bus\Event\EventBus;
 
 final class VideoCreator
