+++ src/Mooc/Videos/Application/Find/FindVideoQueryHandler.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace CodelyTv\Mooc\Videos\Application\Find;
 
-use CodelyTv\Mooc\Videos\Domain\VideoId;
+use CodelyTv\Mooc\Shared\Domain\Videos\VideoId;
 use CodelyTv\Shared\Domain\Bus\Query\QueryHandler;
 use function Lambdish\Phunctional\apply;
 use function Lambdish\Phunctional\pipe;
