+++ src/Mooc/Videos/Application/Find/VideoFinder.php
@@ -4,8 +4,8 @@ declare(strict_types=1);
 
 namespace CodelyTv\Mooc\Videos\Application\Find;
 
+use CodelyTv\Mooc\Shared\Domain\Videos\VideoId;
 use CodelyTv\Mooc\Videos\Domain\VideoFinder as DomainVideoFinder;
-use CodelyTv\Mooc\Videos\Domain\VideoId;
 use CodelyTv\Mooc\Videos\Domain\VideoRepository;
 
 final class VideoFinder
