+++ src/Mooc/Videos/Application/Trim/VideoTrimmer.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace CodelyTv\Mooc\Videos\Application\Trim;
 
-use CodelyTv\Mooc\Videos\Domain\VideoId;
+use CodelyTv\Mooc\Shared\Domain\Videos\VideoId;
 use CodelyTv\Shared\Domain\SecondsInterval;
 
 final class VideoTrimmer
