+++ src/Mooc/Videos/Domain/VideoRepository.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace CodelyTv\Mooc\Videos\Domain;
 
+use CodelyTv\Mooc\Shared\Domain\Videos\VideoId;
 use CodelyTv\Shared\Domain\Criteria\Criteria;
 
 interface VideoRepository
