+++ src/Mooc/Videos/Infrastructure/Persistence/Doctrine/VideoIdType.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace CodelyTv\Mooc\Videos\Infrastructure\Persistence\Doctrine;
 
-use CodelyTv\Mooc\Videos\Domain\VideoId;
+use CodelyTv\Mooc\Shared\Domain\Videos\VideoId;
 use CodelyTv\Shared\Infrastructure\Persistence\Doctrine\UuidType;
 
 final class VideoIdType extends UuidType
