+++ src/Mooc/Videos/Infrastructure/Persistence/Doctrine/VideoUrl.orm.xml
@@ -1,0 +1,11 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<doctrine-mapping xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+                  xmlns="http://doctrine-project.org/schemas/orm/doctrine-mapping"
+                  xsi:schemaLocation="http://doctrine-project.org/schemas/orm/doctrine-mapping
+                          https://www.doctrine-project.org/schemas/orm/doctrine-mapping.xsd">
+
+    <embeddable name="CodelyTv\Mooc\Videos\Domain\VideoUrl">
+        <field name="value" type="string" column="url" />
+    </embeddable>
+
+</doctrine-mapping>
