+++ src/Mooc/Videos/Infrastructure/Persistence/DoctrineVideoRepository.php
@@ -4,8 +4,8 @@ declare(strict_types=1);
 
 namespace CodelyTv\Mooc\Videos\Infrastructure\Persistence;
 
+use CodelyTv\Mooc\Shared\Domain\Videos\VideoId;
 use CodelyTv\Mooc\Videos\Domain\Video;
-use CodelyTv\Mooc\Videos\Domain\VideoId;
 use CodelyTv\Mooc\Videos\Domain\VideoRepository;
 use CodelyTv\Mooc\Videos\Domain\Videos;
 use CodelyTv\Shared\Domain\Criteria\Criteria;
